+++ packages/editor/web/src/RichContentEditor/utils/biUtils.ts
@@ -1,0 +1,19 @@
+export const getBiButtonName = (name: string) => {
+  // eslint-disable-next-line fp/no-loops
+  for (const [key, value] of Object.entries(buttonsNameMap)) {
+    if (value.includes(name)) return key;
+  }
+};
+
+const buttonsNameMap = {
+  Bold: ['bold'],
+  Italic: ['italic'],
+  Underline: ['underline'],
+  OrderedList: ['ordered-list-item'],
+  UnorderedList: ['unordered-list-item'],
+  Blockquote: ['blockquote'],
+  CODE_BLOCK: ['code-block'],
+  LINK: ['link'],
+  DECREASE_INDENT: ['decrease-indent'],
+  INCREASE_INDENT: ['increase-indent'],
+};