+++ packages/ricos-content/web/src/normalization/index.ts
@@ -1,1 +1,2 @@
 export { default as normalizeInitialState } from './normalizeInitialState';
+export { normalizePoll } from './dataNormalizers/pollsDataNormalizer';