+++ packages/ricos-content/web/src/velo-adapter/adapter.ts
@@ -1,2 +1,0 @@
-export { createBuilder } from './builder';
-export { modify } from './modify';