+++ packages/ricos-editor/web/src/RicosEditor.tsx
@@ -344,7 +344,6 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
       theme,
       locale,
       _rcProps: { helpers } = {},
-      toolbarSettings: { getToolbarSettings = () => [] } = {},
       plugins,
       linkPanelSettings,
       linkSettings,
@@ -363,7 +362,8 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
       textToolbarType,
       isMobile,
       theme,
-      getToolbarSettings,
+      toolbarSettings,
+      locale,
       plugins,
       linkPanelSettings,
       linkSettings,
@@ -434,12 +434,10 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
     if (!tiptapEditorModule) {
       return null;
     }
-    const { RicosTiptapEditor, RichContentAdapter } = tiptapEditorModule;
+    const { RicosTiptapEditor, RichContentAdapter, draftToTiptap } = tiptapEditorModule;
     const { content, injectedContent, plugins } = this.props;
     const { tiptapToolbar } = this.state;
-    const initalContent = tiptapEditorModule.draftToTiptap(
-      content ?? injectedContent ?? emptyDraftContent
-    );
+    const initialContent = draftToTiptap(content ?? injectedContent ?? emptyDraftContent);
     const { localeData } = this.state;
     const { locale, localeResource } = localeData;
     const extensions =
@@ -454,7 +452,7 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
               const tiptapEditor = (
                 <RicosTiptapEditor
                   extensions={extensions}
-                  content={initalContent}
+                  content={initialContent}
                   t={t}
                   onLoad={editor => {
                     const richContentAdapter = new RichContentAdapter(editor, t, plugins);