+++ packages/tiptap-editor/web/src/components/BaseComponent/index.ts
@@ -1,1 +1,0 @@
-export { default as BaseExtensionComponentHOC } from './BaseComponent';