+++ packages/tiptap-editor/web/src/models/MarkExtension.ts
@@ -1,7 +1,8 @@
-import { Mark } from '@tiptap/core';
+import { Mark, markInputRule, markPasteRule } from '@tiptap/core';
 import { MarkConfig, mergeAttributes } from '@tiptap/react';
 import { isRicosMarkExtension, RicosExtension } from './extension-types';
 import { DEFAULT_PRIORITY, IMarkExtension } from './domain-types';
+import { Plugin, PluginKey } from 'prosemirror-state';
 
 export class MarkExtension implements IMarkExtension {
   config: MarkConfig;
@@ -18,7 +19,13 @@ export class MarkExtension implements IMarkExtension {
     }
     this.config = {
       addAttributes: () => extension.componentDataDefaults || {},
-      ...extension.createExtensionConfig({ mergeAttributes }),
+      ...extension.createExtensionConfig({
+        mergeAttributes,
+        markInputRule,
+        markPasteRule,
+        Plugin,
+        PluginKey,
+      }),
       type: 'mark',
     };
     this.priority = this.config.priority || DEFAULT_PRIORITY;