+++ packages/toolbars-new/web/src/Toolbar/Toolbar.tsx
@@ -37,7 +37,7 @@ interface ToolbarProps {
 
 interface State {
   buttons: any;
-  overflowedButtons: any;
+  overflowCount: number;
 }
 
 class Toolbar extends Component<ToolbarProps, State> {
@@ -58,12 +58,29 @@ class Toolbar extends Component<ToolbarProps, State> {
       inlineToolbarButton_icon: buttonTheme.textToolbarButton_icon,
     };
     this.theme = { ...props.theme, buttonStyles };
-    this.state = { buttons: props.buttons, overflowedButtons: [] };
+    this.state = { buttons: props.buttons, overflowCount: 0 };
+  }
+
+  componentDidUpdate(prevProps) {
+    if (this.props.buttons !== prevProps.buttons) {
+      const { overflowCount } = this.state;
+      const { isMobile, nestedMenu } = this.props;
+      let { buttons } = this.props;
+      if (!isMobile && this.toolbarRef && !nestedMenu && overflowCount > 0) {
+        buttons = this.updateToolbarOverflow();
+      }
+      // eslint-disable-next-line react/no-did-update-set-state
+      this.setState({ buttons });
+    }
   }
 
   componentDidMount() {
-    const { nestedMenu } = this.props;
-    nestedMenu && this.setFirstAndLastButtons();
+    const { isMobile, nestedMenu } = this.props;
+    let { buttons } = this.state;
+    if (!isMobile && this.toolbarRef && !nestedMenu) {
+      buttons = this.setToolbarOverflow();
+    }
+    this.setState({ buttons }, this.setFirstAndLastButtons);
   }
 
   setFirstAndLastButtons = () => {
@@ -73,20 +90,50 @@ class Toolbar extends Component<ToolbarProps, State> {
     this.lastButton = lastChild?.querySelector?.('button');
   };
 
-  handleToolbarOverflow = () => {
-    const { buttons, overflowedButtons } = this.state;
-    if (this.toolbarOverflowWithEditorWidth()) {
-      if (buttons[buttons.length - 1].name !== 'overflow') {
-        buttons.push({
-          name: 'overflow',
-          type: TOOLBAR_BUTTON_TYPES.NESTED_MENU,
-          getIcon: () => ContextMenuIcon,
-          buttonList: overflowedButtons,
-        });
+  updateToolbarOverflow = () => {
+    const { buttons } = this.props;
+    const { overflowCount } = this.state;
+    const newButtons = [...buttons];
+    const overflowedButtons = newButtons.splice(-overflowCount);
+    newButtons.push({
+      name: 'overflow',
+      type: TOOLBAR_BUTTON_TYPES.NESTED_MENU,
+      getIcon: () => ContextMenuIcon,
+      buttonList: overflowedButtons,
+    });
+    return newButtons;
+  };
+
+  setToolbarOverflow = () => {
+    const { buttons } = this.props;
+    let difference = this.toolbarOverflowWithEditorWidth();
+    const newButtons = [...buttons];
+    const overflowedButtons: any[] = [];
+    if (difference < 0) {
+      const buttonsAsElements = this.toolbarRef.children;
+      let index = buttonsAsElements.length - 1;
+      // eslint-disable-next-line fp/no-loops
+      while (difference < 0) {
+        const currentOverflowedButton = newButtons.splice(-1);
+        overflowedButtons.unshift(...currentOverflowedButton);
+        const lastButtonElement = buttonsAsElements[index] as HTMLElement;
+        difference += lastButtonElement?.clientWidth + 4;
+        index--;
       }
-      const currentOverflowedButton = buttons.splice(-2, 1);
-      overflowedButtons.unshift(...currentOverflowedButton);
-      this.setState({ buttons, overflowedButtons });
+      if (newButtons[newButtons.length - 1].name !== 'Separator') {
+        const currentOverflowedButton = newButtons.splice(-1);
+        overflowedButtons.unshift(...currentOverflowedButton);
+      }
+      newButtons.push({
+        name: 'overflow',
+        type: TOOLBAR_BUTTON_TYPES.NESTED_MENU,
+        getIcon: () => ContextMenuIcon,
+        buttonList: overflowedButtons,
+      });
+      this.setState({ overflowCount: overflowedButtons.length });
+      return newButtons;
+    } else {
+      return buttons;
     }
   };
 
@@ -94,7 +141,7 @@ class Toolbar extends Component<ToolbarProps, State> {
     const rootEditorElement = this.toolbarRef
       ?.closest('[data-hook=ricos-editor-toolbars]')
       ?.parentElement?.querySelector('[data-hook=root-editor]') as HTMLElement;
-    return rootEditorElement.clientWidth < this.toolbarRef.clientWidth;
+    return rootEditorElement?.clientWidth - this.toolbarRef?.clientWidth;
   };
 
   renderButton = buttonProps => {
@@ -243,11 +290,13 @@ class Toolbar extends Component<ToolbarProps, State> {
   renderComponent = buttonProps => {
     const { Component } = buttonProps;
     return (
-      <div className={toolbarButtonStyles.toolbarButton_wrapper}>
-        <button className={toolbarButtonStyles.toolbarButton} onMouseDown={this.preventDefault}>
-          <Component />
-        </button>
-      </div>
+      Component && (
+        <div className={toolbarButtonStyles.toolbarButton_wrapper}>
+          <button className={toolbarButtonStyles.toolbarButton} onMouseDown={this.preventDefault}>
+            <Component />
+          </button>
+        </div>
+      )
     );
   };
 
@@ -305,15 +354,13 @@ class Toolbar extends Component<ToolbarProps, State> {
   setToolbarRef = ref => (this.toolbarRef = ref);
 
   render() {
-    const { isMobile, vertical, nestedMenu } = this.props;
+    const { vertical } = this.props;
     const { buttons } = this.state;
-    !isMobile && this.toolbarRef && !nestedMenu && this.handleToolbarOverflow();
-    this.toolbarRef && !this.toolbarOverflowWithEditorWidth() && this.setFirstAndLastButtons();
 
     // return buttons.map((buttonsWithoutGaps, index) => {
     return (
       <div
-        data-id="toolbar"
+        data-hook="toolbar"
         onKeyDown={this.onKeyDown}
         ref={this.setToolbarRef}
         className={classNames(styles.toolbar, { [styles.vertical]: vertical })}