+++ CHANGELOG.md
@@ -16,29 +16,119 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
- 
-### :bug: Bug Fix
+
+### :house: Internal
+
+- `toolbars-new`
+  - [#3073](https://github.com/wix/ricos/pull/3073) update and add datahooks to the new-toolbar testing
+
+### :nail_care: Polish
 
 - `polls`
+  - [#3026](https://github.com/wix/ricos/pull/3026) refactor settings mobile header / add polls to plugin menu 
+
+### :rocket: New Feature
+
+- `ricos-schema` 
+  - [#3054](https://github.com/wix/ricos/pull/3054) EXTERNAL type
+- `ricos-content`
+  - [#3054](https://github.com/wix/ricos/pull/3054) to/fromDraft: EXTERNAL type supported
+  - [#3056](https://github.com/wix/ricos/pull/3056) EXTERNAL type in EditorCommands
+
+### :bug: Bug Fix
+
+- `toolbar-new`
+  - [#3066](https://github.com/wix/ricos/pull/3066) fix color-picker modal style for mobile devices
+- `toolbars`
+  - [#3063](https://github.com/wix/ricos/pull/3063) ui changes in static text toolbars (old/new toolbar)
+
+### :house: Internal
+
+- `ricos-content`  
+  - [#3071](https://github.com/wix/ricos/pull/3071) Content API: generic modifier and extractor
+
+### :nail_care: Polish
+
+- `polls`
+  - [#3026](https://github.com/wix/ricos/pull/3026) refactor settings mobile header / add polls to plugin menu 
+
+</details>
+<hr/>
+
+## 8.58.3 (Nov 2, 2021)
+
+### :house: Internal
+
+- `tiptap plugins`
+  - [#3047](https://github.com/wix/ricos/pull/3047) add plugins to tiptap [video, gif, file, gallery, code-block, spoiler, color]
+- `video`
+  - [#3050](https://github.com/wix/ricos/pull/3050) open video modal when embed from youtube
+
+## 8.58.2 (Nov 1, 2021)
+
+### :house: Internal
+
+- `schema`
+  - [#3049](https://github.com/wix/ricos/pull/3049) generate-plugin-defaults generates link decoration defaults
+
+### :bug: Bug Fix
+
+- `schema`
+  - add verticala embed schema to content state schemas
+
+## 8.58.1 (Oct 27, 2021)
+
+### :rocket: New Feature
+
+- `social-polls`
+  - [#3041](https://github.com/wix/ricos/pull/3041) fallback user
+
+## 8.58.0 (Oct 27, 2021)
+
+### :rocket: New Feature
+
+- `social-polls`
+  - [#3025](https://github.com/wix/ricos/pull/3025) refactor polls config API & normalize poll data
+- `editor`
+  - [#3028](https://github.com/wix/ricos/pull/3028) plugin menu shortcuts
+
+### :bug: Bug Fix
+
+- `social-polls`
   - [#3020](https://github.com/wix/ricos/pull/3020) fix settings separator style
 - `gallery`
   - [#3033](https://github.com/wix/ricos/pull/3033) fix settings separator style
 
 ### :house: Internal
+
 - `editor-common`
   - [#3030](https://github.com/wix/ricos/pull/3030) refactor getModalStyles (changed modal position when opened from plugin menu)
+- `general`
+  - [#3023](https://github.com/wix/ricos/pull/3023) add BI events to formatting shortcuts
 
-</details>
-<hr/>
+### :book: Documentation
+
+- [#3037](https://github.com/wix/ricos/pull/3037) FromDraftOptions
+
+### :nail_care: Polish
+
+- `ricos-schema`
+  - [#2927](https://github.com/wix/ricos/pull/2927) text decorations and `documentStyle` schema
 
 ## 8.57.0 (Oct 25, 2021)
+
 ### :rocket: New Feature
+
 - `viewer`
   - [#3012](https://github.com/wix/ricos/pull/3012) print mode - fix page break in the middle of list item
+
 ### :bug: Bug Fix
+
 - `ricos-content`
   - [#3019](https://github.com/wix/ricos/pull/3019) fromDraft: Invalid entity key protected
+
 ### :nail_care: Polish
+
 - `ui-components`
   - [#2999](https://github.com/wix/ricos/pull/2999) fix action buttons keyboard navigation order
 - `vertical-embed`
@@ -49,7 +139,9 @@
   - [#3016](https://github.com/wix/ricos/pull/3016) Polls Content Builder API support
 - `ui-components/video/vertical-social-embed`
   - [#3018](https://github.com/wix/ricos/pull/3018) new modals general code / style refactor
+
 ### :house: Internal
+
 - `schema`
   - [#3013](https://github.com/wix/ricos/pull/3013) undefined fields added to plugin defaults JSON
 - `divider`